+++ R/extract_datim.R
@@ -73,7 +73,7 @@ datim_dimension <- function(name,
                               username = accnt$username,
                               password = accnt$password)
 
-  if (base::is.null(df_dims) | base::nrow(df_dims) == 0 | name %ni% df_dims$dimension) {
+  if (base::is.null(df_dims) | base::nrow(df_dims) == 0 | !name %in% df_dims$dimension) {
     base::message(crayon::red(glue::glue("There is no '{name}' dimension. Check spelling.")))
     return(NULL)
   }
@@ -688,7 +688,7 @@ datim_pops <- function(ou,
   # level
   lvl <- level %>% stringr::str_to_lower()
 
-  if (lvl %ni% c("country", "psnu", "prioritization")) {
+  if (!lvl %in% c("country", "psnu", "prioritization")) {
     base::message(crayon::red(glue::glue("Invalid level: {level}")))
     return(NULL)
   }