+++ R/grabr-package.R
@@ -3,7 +3,16 @@
 
 if (getRversion() >= "2.15.1")
   utils::globalVariables(
-    c("value", "Value", "ctime", "filepath")
+    c("%<>%", ".", "age_semi_fine_age", "community_lvl",
+      "convert_datim_pd_to_qtr", "country", "country_iso", "country_lvl",
+      "country_uid", "countryname", "dimension", "disaggregation_type",
+      "displayName", "display_name", "e_tag", "etag", "facility_lvl","filepath",
+      "fiscal_year", "id", "iso3", "iso4", "item", "key", "label",
+      "last_modified", "level", "msd_release", "name3", "name4", "ndir",
+      "object_key", "operatingunit", "operatingunit_iso", "operatingunit_uid",
+      "organisation_unit", "orgunit", "orgunituid", "parent", "path",
+      "prioritization", "psnu_lvl", "regional", "size", "sym", "targets_results",
+      "technical_area", "trendsfine", "type", "uid", "value", "Value", "where")
   )
 
 # The following block is used by usethis to automatically manage